chore: log builder extraction
