+++ logs/log1756115099.md
@@ -1,0 +1,2 @@
+- Moved ConsensusCheckBuilder into its own file to avoid unnecessary nesting and simplify reuse, clarifying the public API per review feedback.
+- Updated gossip and partition code, tests, and docs to reference the standalone builder and ensure consistency.
